NEXT.coin example: fetch BTC, GHOST balance
