+++ examples/next-example/balance.js
@@ -3,10 +3,10 @@ const coins = require('./coins');
 
 (async () => {
 
-/*  const GHOSTbalance = await coins.GHOST.networks.testnet.getBalance('XPtT4tJWyepGAGRF9DR4AhRkJWB3DEBXT2')
-  console.log('GHOST balance: ', GHOSTbalance)*/
-
   let addr, balance
+
+
+  // GHOST
 
   console.log('getBalance: GHOST testnet, zero-balance')
   addr = 'XPtT4tJWyepGAGRF9DR4AhRkJWB3DEBXT2'
@@ -18,13 +18,42 @@ const coins = require('./coins');
   balance = await coins.GHOST.networks.mainnet.getBalance(addr)
   console.log('<<<', balance)
 
-  /*
-  const balanceNEXT = await fetchBalance({
-    coin: coins.NEXT,
-    network: coins.NEXT.networks.mainnet,
-    address: 'XNnXeCcxvPTVFo3DvWERBd6pWZvCfMn9AV'
-  })
-  console.log('GHOST balance: ', balanceGHOST)
-  */
+
+  // BTC
+
+  console.log('getBalance: BTC testnet, zero-balance')
+  addr = '2N3pDTovNkg5QqgkMttjwekPNBZo3m7XfGZ'
+  balance = await coins.BTC.networks.testnet.getBalance(addr)
+  console.log('<<<', balance)
+
+  console.log('getBalance: BTC testnet, non-zero-balance')
+  addr = 'mkHS9ne12qx9pS9VojpwU5xtRd4T7X7ZUt'
+  balance = await coins.BTC.networks.testnet.getBalance(addr)
+  console.log('<<<', balance)
+
+
+  console.log('getBalance: BTC mainnet, zero-balance')
+  addr = '1EPTHasxZ1Hx6tBb8qfLZ3hAEr45y9viU9'
+  balance = await coins.BTC.networks.mainnet.getBalance(addr)
+  console.log('<<<', balance)
+
+  console.log('getBalance: BTC mainnet, non-zero-balance')
+  addr = '3JurbUwpsAPqvUkwLM5CtwnEWrNnUKJNoD'
+  balance = await coins.BTC.networks.mainnet.getBalance(addr)
+  console.log('<<<', balance)
+
+
+/*  // NEXT
+
+  console.log('getBalance: NEXT mainnet, zero-balance')
+  addr = 'XZUsFMpkgPjjfT1c9CwKKyY4TzdQhiKNju'
+  balance = await coins.NEXT.networks.mainnet.getBalance(addr)
+  console.log('<<<', balance)
+
+  console.log('getBalance: NEXT mainnet, non-zero-balance')
+  addr = 'XMkvVVvuQJp4mp5hoVHUPumbnvh63xJsN4'
+  balance = await coins.NEXT.networks.mainnet.getBalance(addr)
+  console.log('<<<', balance)
+*/
 })()
